:sparkles: Feature(webview): manage all pages in one index
