+++ package.json
@@ -386,7 +386,7 @@
     "yarn-audit-fix": "7.0.4"
   },
   "dependencies": {
-    "@luozhu/vscode-channel": "0.6.2",
+    "@luozhu/vscode-channel": "0.7.0",
     "lodash": "^4.17.19",
     "lodash-id": "^0.14.0",
     "picgo": "1.4.25",
