+++ src/vs-picgo/utils/lodash-mixins.ts
