+++ src/webview/message-method.ts
@@ -3,3 +3,4 @@
  * message method from vscode to webview starts with `V2W`
  */
 export const W2V_GET_WEBVIEW_URI = 'W2V_GET_WEBVIEW_URI'
+export const W2V_SHOW_MESSAGE = 'W2V_SHOW_MESSAGE'
